Enterprisify remaining entities
